+++ modules/napsequence/src/sequencecontrollerevent.h
@@ -42,7 +42,7 @@ namespace nap
 		 * @param segmentID the segmentID
 		 * @param time the new time
 		 */
-		double segmentEventStartTimeChange(const std::string& trackID, const std::string& segmentID, float time);
+		double segmentEventStartTimeChange(const std::string& trackID, const std::string& segmentID, double time);
 
 		/**
 		 * overloaded insert segment message
